refactor: extract Unity heartbeat logic to useUnityHeartbeat hook
