+++ src/entities/unity/lib/constants.ts
@@ -1,0 +1,5 @@
+/** Interval between heartbeat Echo messages sent to Unity (ms). */
+export const HEARTBEAT_INTERVAL_MS = 5000;
+
+/** Timeout for a single heartbeat Echo response from Unity (ms). */
+export const HEARTBEAT_TIMEOUT_MS = 3000;
