+++ src/entities/unity/ui/UnityView.tsx
@@ -15,8 +15,8 @@ import { MediaFile } from '@shared/ui/survey/MediaItems/types.ts';
 import {
   useRNUnityCommBridge,
   RNUnityCommBridgeUnityEventHandler,
-  newEchoMessage,
 } from '../lib/hook/useRNUnityCommBridge';
+import { useUnityHeartbeat } from '../lib/hook/useUnityHeartbeat';
 import {
   UnityEventEndUnity,
   UnityEventUnityStarted,
@@ -36,10 +36,10 @@ export const UnityView: FC<Props> = props => {
 
   const rnUnityViewRef = useRef<RNUnityView | null>(null);
   const unityReadyHandled = useRef<boolean>(false);
-  const heartbeatIntervalRef = useRef<ReturnType<typeof setInterval> | null>(null);
   const [unityViewKey, setUnityViewKey] = useState<string | null>(null);
   const { sendMessageToUnity, registerEventHandler, handleMessageFromUnity } =
     useRNUnityCommBridge({ rnUnityViewRef });
+  const { startHeartbeat, stopHeartbeat } = useUnityHeartbeat({ sendMessageToUnity });
   const unityPaths = useRef<Array<string>>([]);
 
   logger.log(`[UnityView]: unityPaths: ${JSON.stringify(unityPaths.current)}`);
@@ -58,37 +58,6 @@ export const UnityView: FC<Props> = props => {
         logger.error(`[UnityView] LoadConfigFile FAILED: ${err}`);
       });
   }, [props.payload.file, logger, sendMessageToUnity]);
-
-  const HEARTBEAT_INTERVAL_MS = 5000;
-  const HEARTBEAT_TIMEOUT_MS = 3000;
-
-  const startHeartbeat = useCallback(() => {
-    if (heartbeatIntervalRef.current) {
-      clearInterval(heartbeatIntervalRef.current);
-    }
-
-    logger.log('[UnityView] Starting periodic heartbeat');
-
-    heartbeatIntervalRef.current = setInterval(() => {
-      const echoPayload = `heartbeat-${Date.now()}`;
-      const echoMsg = newEchoMessage(echoPayload);
-
-      const timeoutId = setTimeout(() => {
-        logger.warn(
-          `[UnityView] Heartbeat Echo timed out after ${HEARTBEAT_TIMEOUT_MS}ms — Unity may be unresponsive`,
-        );
-      }, HEARTBEAT_TIMEOUT_MS);
-
-      sendMessageToUnity(echoMsg)
-        .then(() => {
-          clearTimeout(timeoutId);
-        })
-        .catch(err => {
-          clearTimeout(timeoutId);
-          logger.warn(`[UnityView] Heartbeat Echo failed: ${err}`);
-        });
-    }, HEARTBEAT_INTERVAL_MS);
-  }, [logger, sendMessageToUnity]);
 
   // Register Unity ready handler via the `UnityStarted` event.
   const handleUnityStarted =
@@ -169,10 +138,7 @@ export const UnityView: FC<Props> = props => {
     setUnityViewKey(uuidv4());
 
     return () => {
-      if (heartbeatIntervalRef.current) {
-        clearInterval(heartbeatIntervalRef.current);
-        heartbeatIntervalRef.current = null;
-      }
+      stopHeartbeat();
     };
   }, []);
